Next step : Merge cookie cart and database cart when user connect

+++ resources/js/Cart/HeaderCart/ItemComponent.vue
@@ -1,22 +1,26 @@
 <template>
-    <section class="container">
-        <div class="row align-items-center mb-4 mb-md-2">
-            <div class="col-3 col-md-2">
-                <input :value="quantity" @change="(e) => this.updateItem(e)" class="form-control" type="number">
-            </div>
-            <div class="col-9 col-md-6">
-                {{ item.product_reference.label }}
-            </div>
-            <div class="col-9 col-md-3 text-right">
-                {{ getAmountIncludingTaxes }}
-            </div>
-            <div class="">
-                <button @click="(e) => deleteItem(e)" class="btn btn-sm btn-danger" type="button">
-                    <i class="fas fa-times"></i>
-                </button>
-            </div>
-        </div>
-    </section>
+    <tr class="header-cart-item">
+        <td>
+            <input :value="quantity" @change="(e) => this.updateItem(e)" class="form-control header-cart-item-quantity"
+                   min="1" type="number">
+        </td>
+        <td>
+            <a :href="item.product_reference.url">{{ item.product_reference.label }}</a>
+        </td>
+        <td>
+            {{ getAmountIncludingTaxes }}
+        </td>
+        <td class="header-cart-item-td-icon">
+            <button @click="(e) => deleteItem(e)" class="btn btn-sm btn-link text-danger p-0" type="button"
+                    v-show="!isLoading">
+                <i class="fas fa-times"></i>
+            </button>
+
+            <span v-show="isLoading">
+                <i class="fas fa-spinner spinner"></i>
+            </span>
+        </td>
+    </tr>
 </template>
 
 
@@ -38,17 +42,24 @@
         computed: {
             getAmountIncludingTaxes: function () {
                 return this.toLocaleCurrency(this.item.amount_including_taxes, CartStore.data.currency);
+            },
+            isLoading: function () {
+                return CartStore.state.isLoading[this.item.product_reference_id];
             }
         },
         methods: {
-            deleteItem: function (event) {
-                CartStore.deleteItem(this.item.product_reference_id);
-                event.stopPropagation();
+            deleteItem: function () {
+                if (!this.isLoading) {
+                    CartStore.deleteItem(this.item.product_reference_id);
+                }
             },
             updateItem: function (event) {
-                CartStore.updateItem(this.item.product_reference_id, event.target.value);
-                this.quantity = event.target.value;
-                event.stopPropagation();
+                const quantity = parseFloat(event.target.value);
+
+                if (quantity > 0) {
+                    CartStore.updateItem(this.item.product_reference_id, quantity);
+                    this.quantity = quantity;
+                }
             }
         }
     }
